requirements.txt
adrf==0.1.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.10
async-property==0.2.2
attrs==23.2.0
autobahn==24.4.2
Automat==25.4.16
bandit==1.8.3
beautifulsoup4==4.13.4
bidict==0.23.1
black==24.10.0
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
ConfigArgParse==1.7.1
constantly==23.10.4
crashtest==0.4.1
cryptography==45.0.3
daphne==4.2.1
dill==0.4.0
distlib==0.3.9
Django==5.0.7
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==24.3
django-silk==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2025.4.1
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.1.1
flask-cors==6.0.0
Flask-Login==0.6.3
frozenlist==1.7.0
gevent==25.5.1
geventhttpclient==2.3.3
gprof2dot==2025.4.14
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
identify==2.6.12
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
installer==0.7.0
isort==6.0.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==25.6.0
locust==2.37.10
locust-cloud==1.23.1
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.0
multidict==6.6.4
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pbr==6.1.1
pbs-installer==2025.6.6
pillow==10.4.0
pip-tools==7.5.0
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.5.0
poetry==2.1.3
poetry-core==2.1.3
pre-commit==3.8.0
propcache==0.3.2
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.8.0
pylint==3.3.7
pyOpenSSL==25.1.0
pyproject_hooks==1.2.0
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.28.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-slugify==8.0.4
python-socketio==5.13.0
python-telegram-bot==22.3
pytz==2025.2
PyYAML==6.0.1
pyzmq==26.4.0
RapidFuzz==3.13.0
redis==6.4.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.19.0
ruff==0.5.7
SecretStorage==3.3.3
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stevedore==5.4.1
text-unidecode==1.3
tomlkit==0.13.2
trove-classifiers==2025.5.9.12
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.31.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yarl==1.20.1
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
